Create README.md
Offered without guarantee or warranty or any of that stuff; eventually I'll put a license file here. Some of these may not work with current versions of Python/AstroPy/etc. If you know of better ways to implement any of these various tasks I'd be happy to hear about them.